+++ src/spec/__mocks__/amazon-cognito-identity-js/CognitoUserPool.ts
@@ -1,0 +1,9 @@
+export class CognitoUserPool {
+  public userPoolId = '';
+  public clientId = '';
+  public getCurrentUser = jest.fn().mockReturnValue('cognitouserpool');
+
+  constructor(data: Partial<CognitoUserPool>) {
+    Object.assign(this, data);
+  }
+}