+++ docs/changelog.md
@@ -11,6 +11,7 @@ Sprint 11 / April 26, 2020
 - added bar graphs to Graph component.
 - updated theme and global styles.
 - animated Spinner component.
+- added AWS Cognito Identity SDK mock for testing auth flow.
 
 <hr />